:sparkles: experimental framework for static sparse feedforward nns done

+++ StaticConfigurationGenerator.py
@@ -43,20 +43,28 @@ ratios = [1.00, 0.90, 0.80, 0.70, 0.60, 0.50, 0.40, 0.30, 0.20, 0.10, 0.00]
 max_connection_depths = [1, 2, 3, 4]
 
 for dataset in datasets:
+    if not os.path.isdir(f"experiments/static/{dataset}/"):
+        os.mkdir(f"experiments/static/{dataset}/")
     for sparsity in sparsities:
         for ratio in ratios:
             for max_connection_depth in max_connection_depths:
+                _ratio = ratio
                 config = base_configuration.copy()
                 config["trainer_config"]["dataset"] = dataset
                 config["model_config"]["max_connection_depth"] = max_connection_depth
                 config["model_config"]["sparsity"] = sparsity
-                config["model_config"]["skip_sequential_ratio"] = ratio
+                if max_connection_depth == 1:
+                    _ratio = 1
+                config["model_config"]["skip_sequential_ratio"] = _ratio
 
-                directory_name = f"DS-{dataset}_MCD-{max_connection_depth}_S-{sparsity}_R-{ratio}"
+                directory_name = f"DS-{dataset}_MCD-{max_connection_depth}_S-{sparsity}_R-{_ratio}"
 
-                os.mkdir(f"experiments/static/{dataset}/{directory_name}")
+                if not os.path.isdir(f"experiments/static/{dataset}/{directory_name}"):
+                    os.mkdir(f"experiments/static/{dataset}/{directory_name}")
 
-                with open(f"experiments/static/{dataset}/{directory_name}/config.pkl", "wb") as file:
-                    pkl.dump(config, file)
+                if not os.path.exists(f"experiments/static/{dataset}/{directory_name}/config.pkl"):
+                    print(f"creating experiments/static/{dataset}/{directory_name}/config.pkl")
+                    with open(f"experiments/static/{dataset}/{directory_name}/config.pkl", "wb") as file:
+                        pkl.dump(config, file)
